fix(chat-deploy): fixed form submission access patterns
