+++ apps/sim/app/workspace/[workspaceId]/w/[workflowId]/components/control-bar/components/deploy-modal/components/chat-deploy/hooks/index.ts
@@ -1,0 +1,6 @@
+import { useChatDeployment } from './use-chat-deployment'
+import { useChatForm } from './use-chat-form'
+
+export { useChatDeployment, useChatForm }
+
+export type { ChatFormData, ChatFormErrors } from './use-chat-form'
